Merged RSA and ECDSA key generation into PV204-keygen in pv204
Now all key generation is done in a single program - it is more effective as generation of keys of both algorithms uses the same functions for time measurement and RNG seeding. This means improvement of RNG seeding for RSA keys generation - the function from ecdsa-keygen is better, and therefore used in PV204-keygen. Also simplified error handling in time measurement functions (it is also now similar as it was in ecdsa-keygen) and fixed missing file closure at the end of ECDSA key generation.